fix todo: (#4149)
1. only support convert integer string to integer
2.  fix fogInfo sevalue_to_native of jsb_conversions_spec.cpp
3. fix ShadowsInfo sevalue_to_native of jsb_conversions_spec.cpp
4. add leak field of FogInfo & Shadows at scene.ini
5. fix todo of SkeletalAnimationUtils.h

+++ native/cocos/3d/skeletal-animation/SkeletalAnimationUtils.h
@@ -40,9 +40,10 @@ namespace cc {
 
 class Mesh;
 
+// _chunkIdxMap[key] = skeleton ^ clips[i]
 struct IChunkContent {
-    int32_t              skeleton{0}; // TODO(xwx): int or uint or float?
-    std::vector<int32_t> clips;       // TODO(xwx): int or uint?
+    uint64_t              skeleton{0};
+    std::vector<uint64_t> clips;
 };
 
 struct ICustomJointTextureLayout {
